fix more format specifiers for printing F77_INT variables

+++ src/sl_ab13bd.cc
@@ -125,7 +125,7 @@ For internal use only.")
             error ("lti: norm: __sl_ab13bd__: AB13BD returned info = %d", static_cast<int> (info));
 
         if (iwarn != 0)
-            warning ("lti: norm: __sl_ab13bd__: AB13BD returned iwarn = %d", iwarn);
+            warning ("lti: norm: __sl_ab13bd__: AB13BD returned iwarn = %d", static_cast<int> (iwarn));
         
         // return value
         retval(0) = octave_value (norm);
